project/requirements.txt
pillow
django-toolbelt
gunicorn
